+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
+## 0.11.0 - 2024-11-04
 ### Added
 - `duration`:
     - Wrapper for importing Duration class in the same way as `rclpy`.